added "@" toggle transparency token.
This new token allows patterns to be transparent.
I.e. the color from the pattern beneath the currently active one is displayed.
This works for any number of levels, the first non transparent pattern is
used, or the lowest pattern if all are transparent.

The transparency of the lowest pattern is currently ignored.