fixed implementation of  brightness
with the previous implementation, when brightness reached 0 the color
was lost, and when the brightness was increased or decreased to the
extremes the color could change slightly.

This new implementation fixes this by keeping the color as a base
color. And then adjusting the brightness with a separate value.

This preserves the exact color value even at 0 brightness.