added 8 to 12 bit led brightness adjustment
with the linear mapping from 8 to 12 bits, the lowest brightness of
value 1 was already mapped to 16 which because of the non linearity
of LED's is already very bright, on the other hand the difference at
the upper end was negligible.

This logarithmic mapping decreases the step size in the low end and
increases it in the high end. This allows for much finer brightness control.

TODO: fix the reverse
Ideally rgb8torgb12 would be the exact inverse of rgb12torgb8, this is
currently not the case. rgb12torgb8 still uses the pure linear
mapping. A binary search approach could probably be used to utilize
the same log_map.